Share Network enum in both binaries
Remove unnecessary duplication to ease maintenance
