+++ shared-bin/Cargo.toml
@@ -10,6 +10,8 @@ description = "Code that is shared between the daemons but application specific 
 [dependencies]
 anyhow = "1"
 atty = "0.2"
+bitmex-stream = { path = "../bitmex-stream" }
+clap = { version = "3", features = ["derive"] }
 console-subscriber = "0.1.6"
 daemon = { path = "../daemon" }
 http-api-problem = { version = "0.53.0", features = ["rocket"] }
